+++ CHANGELOG.md
@@ -7,13 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Introduced the concept of the `Workspace`
+- Added a stub of a webserver that will be able to visualize runs as they happen.
+
 ### Changed
 
 - Made it possible to construct a step graph out of `Step` objects, instead of constructing it out of `StepStub` objects.
 - Removed dataset fingerprinting code, since we can now use `Step` to make sure things are cached.
-- Made `Executor` non-registrable. This is a temporary state and will be changed back.
 - Made steps deterministic by default.
 - Brought back `MemoryStepCache`, so we can run steps without configuring anything.
+- W&B `torch::TrainCallback` logs with `step=step+1` now so that training curves in the W&B dashboard
+  match up with checkpoints saved locally and are easier to read (e.g. step 10000 instead of 9999).
 
 ## [v0.3.6](https://github.com/allenai/tango/releases/tag/v0.3.6) - 2021-11-12