+++ CHANGELOG.md
@@ -7,13 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Introduced the concept of the `Workspace`, with `LocalWorkspace` and `MemoryWorkspace` as initial implementations.
+- Added a stub of a webserver that will be able to visualize runs as they happen.
+- Added separate classes for `LightningTrainingTypePlugin`, `LightningPrecisionPlugin`, `LightningClusterEnvironmentPlugin`, `LightningCheckpointPlugin` for compatibility with `pytorch-lightning>=1.5.0`.
+
+### Removed
+
+- Removed old `LightningPlugin` class.
+
 ### Changed
 
 - Made it possible to construct a step graph out of `Step` objects, instead of constructing it out of `StepStub` objects.
 - Removed dataset fingerprinting code, since we can now use `Step` to make sure things are cached.
-- Made `Executor` non-registrable. This is a temporary state and will be changed back.
 - Made steps deterministic by default.
 - Brought back `MemoryStepCache`, so we can run steps without configuring anything.
+- W&B `torch::TrainCallback` logs with `step=step+1` now so that training curves in the W&B dashboard
+  match up with checkpoints saved locally and are easier to read (e.g. step 10000 instead of 9999).
+
+### Fixed
+
+- Fixed bug in `FromParams` where a parameter to a `FromParams` class may not be instantiated correctly
+  if it's a class with a generic type parameter.
 
 ## [v0.3.6](https://github.com/allenai/tango/releases/tag/v0.3.6) - 2021-11-12