+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Brought back `MemoryStepCache`, so we can run steps without configuring anything.
 - W&B `torch::TrainCallback` logs with `step=step+1` now so that training curves in the W&B dashboard
   match up with checkpoints saved locally and are easier to read (e.g. step 10000 instead of 9999).
+- `filelock >= 3.4` required, parameter `poll_intervall`  to `tango.common.file_lock.FileLock.acquire` renamed
+  to `poll_interval`.
 
 ### Fixed